支持各协议发送路径请求；新增注解式业务层。

+++ src/main/java/cn/messageplus/core/request/MessageRequest.java
@@ -17,5 +17,5 @@ public abstract class MessageRequest implements Serializable {
     /**
      * 自定义请求类型字节码需要大于等于100
      */
-    public abstract byte getType();
+    public abstract short getType();
 }
