+++ src/main/java/cn/messageplus/core/request/RequestFactory.java
@@ -9,19 +9,19 @@ import java.util.Map;
  **/
 public class RequestFactory {
     // 存储请求类型及请求类型的代码
-    public static Map<Integer, Class<? extends MessageRequest>> messageClasses = new HashMap<>();
+    public static Map<Short, Class<? extends MessageRequest>> messageClasses = new HashMap<>();
 
     /**
      * 添加请求类型
      */
     public static void addRequestType(List<MessageRequest> requestTypes) {
         for (MessageRequest requestType : requestTypes) {
-            messageClasses.put((int) requestType.getType(), requestType.getClass());
+            messageClasses.put(requestType.getType(), requestType.getClass());
         }
     }
 
-    public static Class<? extends MessageRequest> getRequestType(byte type) {
-        return messageClasses.get((int) type);
+    public static Class<? extends MessageRequest> getRequestType(short type) {
+        return messageClasses.get(type);
     }
 
 }
